backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
python-jose[cryptography]
python-multipart
httpx
requests
pydantic